+++ changes.md
@@ -1,8 +1,18 @@
-**Import changes**
+**Api changes**
 
 <details>
-<summary>Removed Type(s)</summary>
+<summary>Added Resource(s)</summary>
 
-- :warning: removed type `ReferencedResourceNotFound`
+- added resource `/{projectKey}/channels/key={key}`
+</details>
+
+
+<details>
+<summary>Added Method(s)</summary>
+
+- added method `apiRoot.withProjectKey().channels().withKey().get()`
+- added method `apiRoot.withProjectKey().channels().withKey().head()`
+- added method `apiRoot.withProjectKey().channels().withKey().post()`
+- added method `apiRoot.withProjectKey().channels().withKey().delete()`
 </details>